refactor/GRAN MODULARIZACIÓN de todas las rutas
